Moved files to appropriate placements
