+++ app/components/HomeTabs/index.js
