+++ app/index.js
@@ -7,14 +7,14 @@ import { Provider } from 'react-redux';
 import { createStore, combineReducers } from 'redux';
 
 // reducers
-import * as allReducers from './app/redux/reducers';
+import * as allReducers from './reducers';
 
 // ui
-import App from './App';
+import App from './containers/App';
 
 const store = createStore(combineReducers(allReducers));
 
-export default class Home extends Component {
+export default class GTHC extends Component {
   render() {
     return (
       <Provider store={store}>
