+++ app/reducers/posts.js
