+++ app/reducers/shifts.js
