+++ index.js
@@ -1,8 +1,6 @@
 /** @format */
-
 import { AppRegistry } from 'react-native';
-import Home from './Home';
-import Login from './Login';
 
-AppRegistry.registerComponent('GTHC', () => Home);
-AppRegistry.registerComponent('Login', () => Login);
+import GTHC from './app';
+
+AppRegistry.registerComponent('GTHC', () => GTHC);
